Move global layout control to layout.py
